pack-format: document missing fields
Document missing fields in the description of the delta data, added at
011b648 (pack-format.txt: more details on pack file format, 2018-05-11).

Also, the description of object entries at the main section is a bit
vague. There's an equivalent but more detailed description in the index
section; since they are redundant, put that one in place of the first.

I have also expanded tabs to prevent alignment issues and rephrased a
bit to make things clearer.

Signed-off-by: Alba Mendez <me@alba.sh>