git-p4.py: fix --prepare-p4-only error with multiple commits
When using git p4 submit with the --prepare-p4-only option, the program
should prepare a single p4 changelist and notify the user that more
commits are pending and then stop processing.

A bug has been introduced by the p4-changelist hook feature that
causes the program to continue to try and process all pending
changelists at the same time.

The function applyCommit should return True when applying the commit
was successful and the program should continue. In the case of the
--prepare-p4-only flag, the function should return False, alerting the
caller that the program should not proceed with additional commits.

Change the return value from True to False in the applyCommit function
when git-p4 is executed with --prepare-p4-only flag.

Signed-off-by: Ben Keene <seraphire@gmail.com>