Makefile: skip git-gui if dir is missing
As an effort to promote "dogfooding" the sparse-checkout feature within
the Git codebase, it is helpful to explore which portions of the
codebase are optional.

The NO_TCLTK variable in the Makefile allows ignoring the git-gui
application at build time. If the local sparse-checkout removes that
directory, then the build will fail without manually defining that
variable. Instead, check to see if the directory exists at build time to
see if we can automatically ignore this directory.

With this change, the following list of directories can be supplied to
"git sparse-checkout set" and Git will build and test on Linux:

	Documentation
	builtin
	compat/.depend
	contrib
	ewah
	mergetools
	negotiator
	perl
	refs
	sha1dc
	sha256
	t
	templates
	trace2
	vcs-svn
	xdiff

The "make install" command requires the "gitweb" and "po" directories.
The "po" directory will become optional in the next change.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>